northstar-agents/pyproject.toml
[project]
name = "northstar-agents"
version = "0.1.0"
description = "Assembly line for agents"
authors = [{name = "Northstar Team"}]
requires-python = ">=3.11"
dependencies = [
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML",
    "pre-commit",
]
bedrock = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "types-boto3",
    "botocore-stubs",
]
adk = [
    "google-auth>=2.0.0",
]
autogen = [
    "pyautogen>=0.2.0",
]
crewai = [
    "crewai>=0.1.0",
]
langgraph = [
    "langgraph>=0.0.1",
]
strands = [
    "strands>=0.0.1",
]
azure = [
    "azure-identity>=1.15.0",
    "azure-ai-openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
northstar = "northstar.__main__:run_cli"

[tool.hatch.build.targets.wheel]
packages = ["src/northstar"]

[tool.ruff]
target-version = "py311"

[tool.mypy]
strict = true
ignore_missing_imports = true
